// frontend/src/rebuild/learning/horizonte/com/trig.generated.ts
export const LEVELS = ['zero', 'half', 'root2', 'root3', 'one'] as const;
export type Level = (typeof LEVELS)[number];